refactor: CardShape과 CardNumber를 관리하는 CardProperty 추가
