+++ src/main/java/blackjack/domain/card/CardGroup.java
@@ -1,5 +1,6 @@
 package blackjack.domain.card;
 
+import blackjack.domain.card.property.CardNumber;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
