+++ src/main/java/blackjack/domain/card/property/CardProperty.java
@@ -1,0 +1,23 @@
+package blackjack.domain.card.property;
+
+public class CardProperty {
+    private final CardShape shape;
+    private final CardNumber number;
+
+    public CardProperty(CardShape shape, CardNumber number) {
+        this.shape = shape;
+        this.number = number;
+    }
+
+    public CardShape getShape() {
+        return shape;
+    }
+
+    public CardNumber getNumber() {
+        return number;
+    }
+
+    public boolean isA() {
+        return number.isA();
+    }
+}
